+++ Fabric/src/main/java/fuzs/moblassos/fabric/world/item/FabricLassoItem.java
@@ -1,7 +1,7 @@
 package fuzs.moblassos.fabric.world.item;
 
-import fuzs.moblassos.world.item.LassoItem;
-import fuzs.moblassos.world.item.LassoType;
+import fuzs.moblassos.common.world.item.LassoItem;
+import fuzs.moblassos.common.world.item.LassoType;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;