+++ NeoForge/src/main/java/fuzs/moblassos/neoforge/init/NeoForgeModRegistry.java
@@ -1,9 +1,9 @@
 package fuzs.moblassos.neoforge.init;
 
-import fuzs.moblassos.MobLassos;
-import fuzs.moblassos.init.ModRegistry;
+import fuzs.moblassos.common.MobLassos;
+import fuzs.moblassos.common.init.ModRegistry;
 import fuzs.moblassos.neoforge.world.item.NeoForgeLassoItem;
-import fuzs.moblassos.world.item.LassoType;
+import fuzs.moblassos.common.world.item.LassoType;
 import fuzs.puzzleslib.common.api.init.v3.registry.RegistryManager;
 import net.minecraft.core.Holder;
 import net.minecraft.world.item.Item;